requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
databricks-sql-connector==4.2.3
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.127.0
h11==0.16.0
idna==3.11
lz4==4.4.5
numpy==2.0.2
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
pyarrow==21.0.0
pybreaker==1.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
starlette==0.49.3
thrift==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.39.0
